Changing some lingering uses of komi as an int

+++ include/budgie/budgie.hpp
@@ -96,7 +96,7 @@ class budgie {
 
 		args_parser::option_map options;
 
-		int komi;
+		float komi;
 		int boardsize;
 		int playouts;
 		bool passed = false;
